+++ example/lib/main.dart
@@ -1,8 +1,6 @@
-import 'package:example/bindings/bindings.dart';
+import 'package:example/conditional_navigation_example.dart';
 import 'package:flutter/material.dart';
 import 'package:get_x_master/get_x_master.dart';
-
-import 'controller/theme_controller.dart';
 /*
 void main() {
   runApp(const MyApp());
@@ -32,59 +30,5 @@ class MyApp extends StatelessWidget {
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  runApp(const ColorStudioApp());
-}
-
-class ColorStudioApp extends StatelessWidget {
-  const ColorStudioApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return GetMaterialApp(
-      debugShowCheckedModeBanner: false,
-      initialBinding: MyBinding(),
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(
-          seedColor: Colors.deepOrange,
-          brightness: Brightness.light,
-        ),
-        useMaterial3: true,
-      ),
-      darkTheme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(
-          seedColor: Colors.deepOrange,
-          brightness: Brightness.dark,
-        ),
-        useMaterial3: true,
-      ),
-      home: HomeScreen(),
-    );
-  }
-}
-
-class HomeScreen extends StatelessWidget {
-  const HomeScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('تغییر تم')),
-      body: Center(
-        child: GetBuilder<ThemeController>(
-          init: null, // Let GetX find the controller from bindings
-          builder: (controller) {
-            return Obx(
-              () => Switch(
-                value: controller.isDarkMode,
-                onChanged: (value) {
-                  controller.toggleTheme();
-                },
-                activeThumbColor: Colors.deepOrange,
-              ),
-            );
-          },
-        ),
-      ),
-    );
-  }
+  runApp(const GetMaterialApp(home: StartPage()));
 }